backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
python-multipart
python-dotenv
pydantic
